dds service calls; generic plugins; new idls;
- add navigation reference, arm state, button event, operational mode
- add service calls, add generic plugins, revise wrapper for enums

+++ modules/core_framework/src/service_layer.cpp
@@ -8,7 +8,7 @@
 using namespace dls;
 
 
-ServiceLayer::ServiceLayer(std::string ID, const std::string& robot_name) 
+ServiceLayer::ServiceLayer(std::string ID, const std::string& robot_name)
 	: Layer(ID, 500)
 	, ddsMonitor(std::make_shared<dls::DDSWriter>(
 		"ServiceLayer::monitor",
@@ -58,7 +58,7 @@ ServiceLayer::ServiceLayer(std::string ID, const std::string& robot_name)
 		{{0,1},{1,1}},
 		true
 	);
-	
+
 	command_manager.addCommand<>
 	(
 		"unloadDataVisualizer",
@@ -106,6 +106,35 @@ ServiceLayer::ServiceLayer(std::string ID, const std::string& robot_name)
 		true
 	);
 
+  command_manager.addCommand<std::string>
+  (
+    "loadGeneric",
+    "Load a generic",
+    std::function<bool(std::string)>([&](std::string type)->bool
+        {
+      return this->loadGeneric(type);
+        }),
+    {},
+    true
+  );
+
+  command_manager.addCommand<std::string>
+  (
+    "unloadGeneric",
+    "Remove generic",
+    std::function<bool(std::string)>([&](std::string type)->bool
+        {
+      if(this->unloadGeneric(type))
+      {
+        return true;
+      }
+
+      return false;
+    }),
+    {},
+    true
+  );
+
 	command_manager.addCommand<std::string>
 	(
 		"loadTask",
@@ -148,13 +177,15 @@ void ServiceLayer::close(){
 		this->unloadService(pair.first);
 	for(auto pair : this->data_visualizers_)
 		this->unloadDataVisualizer(pair.first);
+	for(auto pair : this->generics)
+		this->unloadGeneric(pair.first);
 	for(auto pair : this->tasks)
 		this->unloadTask(pair.first);
 }
 
 bool ServiceLayer::loadService(const std::string& lib_name)
 {
-	
+
 	if(this->services.find(lib_name) != this->services.end())
 	{
 		scout_err << "service " + lib_name + " already loaded" << std::endl;
@@ -162,7 +193,7 @@ bool ServiceLayer::loadService(const std::string& lib_name)
 	}
 
     std::shared_ptr<AppData> pData = std::make_shared<AppData>(lib_name);
-    
+
 	{
 		// std::lock_guard<std::mutex> lock(this->services_mutex);
 
@@ -222,7 +253,7 @@ bool ServiceLayer::unloadService(const std::string ID)
 			return true;
 		}), 2000, 10, unloaded)){
 		scout_warn << "### FORCING SERVICE " << ID << " EXIT ###" << std::endl;
-		kill(pData->proc->id(), SIGKILL);		
+		kill(pData->proc->id(), SIGKILL);
 	}
 
 	scout_sys << "Service " + ID + " is unloaded" << std::endl;
@@ -240,7 +271,7 @@ bool ServiceLayer::loadDataVisualizer(const std::string& lib_name)
 	}
 
     std::shared_ptr<AppData> pData = std::make_shared<AppData>(lib_name);
-    
+
 	{
 		// std::lock_guard<std::mutex> lock(this->data_visualizers_mutex_);
 
@@ -298,12 +329,85 @@ bool ServiceLayer::unloadDataVisualizer(const std::string ID)
 			return true;
 		}), 2000, 10, unloaded)){
 		scout_warn << "### FORCING DATA VISUALIZER " << ID << " EXIT ###" << std::endl;
-		kill(pData->proc->id(), SIGKILL);		
+		kill(pData->proc->id(), SIGKILL);
 	}
 
 	scout_sys << "Data visualizer " + ID + " is unloaded" << std::endl;
 	pData->proc = nullptr;
 	this->data_visualizers_.erase(ID);
+    return true;
+}
+
+bool ServiceLayer::loadGeneric(const std::string& name)
+{
+  if (this->generics.find(name) != this->generics.end())
+  {
+    scout_err << "generic " + name + " already loaded" << std::endl;
+    return false;
+  }
+
+  std::shared_ptr<AppData> pData = std::make_shared<AppData>(name);
+
+  // launch the generic
+  char *child_process_launcher = std::getenv("DLS_CHILD_PROCESS_LAUNCHER");
+  if (!child_process_launcher)
+  {
+    scout_err <<
+      "env variable DLS_CHILD_PROCESS_LAUNCHER not "
+      "defined.  This is probably an error with the launch script"
+      << std::endl;
+    return false;
+  }
+
+  pData->proc = std::make_shared<boost::process::child>(std::vector<std::string>({
+    child_process_launcher,
+    pData->getID(),
+    name,
+    "generic",
+    this->robot_name
+  }));
+
+  if (pData->proc == nullptr)
+  {
+    std::cout << "Generic " << name <<" failed to launch: nullptr" << std::endl;
+    return false;
+  }
+
+  pData->proc->detach();
+
+  this->generics.emplace(pData->getID(), pData);
+
+  return true;
+}
+
+bool ServiceLayer::unloadGeneric(const std::string &name)
+{
+  // Find generic inside the generic list
+  auto res = this->generics.find(name);
+
+  if (res == this->generics.end())
+  {
+    scout_err << "Generic " + name + " is not loaded" << std::endl;
+    return false;
+  }
+
+  auto pData = res->second;
+
+    //shutdown generic over the dds comunication layer
+  command_manager.callCommand("shutdown", {}, name);
+
+  bool unloaded = false;
+  if (!utils::wait(std::function<bool()>([&](){ return !pData->proc->running(); }),
+                   2000, 10, unloaded))
+  {
+    std::cout << "### FORCING GENERIC " << name << " EXIT ###" << std::endl;
+    kill(pData->proc->id(), SIGKILL);
+  }
+
+  std::cout << "Generic " + name + " is unloaded" << std::endl;
+
+  pData->proc = nullptr;
+  this->generics.erase(name);
     return true;
 }
 
@@ -316,7 +420,7 @@ bool ServiceLayer::loadTask(const std::string& name)
 	}
 
     std::shared_ptr<AppData> pData = std::make_shared<AppData>(name);
-    
+
 	// launch the procedure
 	char *child_process_launcher = std::getenv("DLS_CHILD_PROCESS_LAUNCHER");
 	if(!child_process_launcher)
@@ -340,12 +444,12 @@ bool ServiceLayer::loadTask(const std::string& name)
 		std::cout << "Task " << name <<" failed to launch: nullptr" << std::endl;
 		return false;
 	}
-	
+
 	pData->proc->detach();
-	
+
 	this->tasks.emplace(pData->getID(), pData);
 
-	return true;	
+	return true;
 }
 
 bool ServiceLayer::unloadTask(const std::string &name)
@@ -372,11 +476,11 @@ bool ServiceLayer::unloadTask(const std::string &name)
 			return true;
 		}), 2000, 10, unloaded)){
 		std::cout << "### FORCING TASK " << name << " EXIT ###" << std::endl;
-		kill(pData->proc->id(), SIGKILL);		
+		kill(pData->proc->id(), SIGKILL);
 	}
-	
+
 	std::cout << "Task " + name + " is unloaded" << std::endl;
-	
+
 	pData->proc = nullptr;
 	this->tasks.erase(name);
     return true;
